01-tutorials/02-AgentCore-gateway/04-bearer-token-injection/requirements.txt
strands-agents
strands-agents-tools
bedrock-agentcore
PyYAML
ddgs
bedrock_agentcore_starter_toolkit
requests